Rename components that conflict with HTML elements
