+++ stubs/inertia/resources/js/Components/ConfirmsPassword.vue
@@ -1,10 +1,10 @@
 <script setup>
 import { ref, reactive, nextTick } from 'vue';
 import DialogModal from './DialogModal.vue';
-import Input from './Input.vue';
 import InputError from './InputError.vue';
 import PrimaryButton from './PrimaryButton.vue';
 import SecondaryButton from './SecondaryButton.vue';
+import TextInput from './TextInput.vue';
 
 const emit = defineEmits(['confirmed']);
 
@@ -85,7 +85,7 @@ const closeModal = () => {
                 {{ content }}
 
                 <div class="mt-4">
-                    <Input
+                    <TextInput
                         ref="passwordInput"
                         v-model="form.password"
                         type="password"
