+++ stubs/inertia/resources/js/Components/InputLabel.vue
