+++ stubs/inertia/resources/js/Components/TextInput.vue
