+++ stubs/inertia/resources/js/Pages/Auth/ConfirmPassword.vue
@@ -3,10 +3,10 @@ import { ref } from 'vue';
 import { Head, useForm } from '@inertiajs/inertia-vue3';
 import AuthenticationCard from '@/Components/AuthenticationCard.vue';
 import AuthenticationCardLogo from '@/Components/AuthenticationCardLogo.vue';
-import Input from '@/Components/Input.vue';
 import InputError from '@/Components/InputError.vue';
-import Label from '@/Components/Label.vue';
+import InputLabel from '@/Components/InputLabel.vue';
 import PrimaryButton from '@/Components/PrimaryButton.vue';
+import TextInput from '@/Components/TextInput.vue';
 
 const form = useForm({
     password: '',
@@ -39,8 +39,8 @@ const submit = () => {
 
         <form @submit.prevent="submit">
             <div>
-                <Label for="password" value="Password" />
-                <Input
+                <InputLabel for="password" value="Password" />
+                <TextInput
                     id="password"
                     ref="passwordInput"
                     v-model="form.password"
